flood-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
earthengine-api==0.1.395
xgboost==2.0.3
scikit-learn==1.4.2
numpy==1.26.4
shapely==2.0.4
geojson==3.1.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
